refactor(dataset): the episode ground-truth reader lives with its verifier
read_dataset_episode_indices moves from dataset_recorder (which never
called it) into verify_dataset (whose whole job it serves); the three
callers and every cite follow. No behaviour change.
