+++ strands_robots/utils.py
@@ -1375,7 +1375,7 @@ def declared_count(value: object) -> int | None:
     answers a reader can act on are the count itself and the absence of one.
     Every reader of a LeRobot header count asks that question of the same file -
     the parquet cross-check in
-    :func:`~strands_robots.dataset_recorder.read_dataset_episode_indices`, the
+    :func:`~strands_robots.verify_dataset.read_dataset_episode_indices`, the
     metadata-drift check in
     :func:`~strands_robots.verify_dataset.verify_dataset`, the validation-split
     denominator in ``strands_robots.training.lerobot``, the episode count the
